Move functions, rename hamiltonian.py
